+++ backend/app/tasks/__init__.py
@@ -1,0 +1,7 @@
+"""
+Celery tasks package for video transcription
+"""
+
+from app.tasks.transcription import transcribe_video
+
+__all__ = ["transcribe_video"]
