--- backend/app/utils/encryption.py
@@ -1,43 +1,0 @@
-"""Utility functions for encryption and decryption"""
-
-import base64
-
-from cryptography.fernet import Fernet
-from django.conf import settings
-
-
-class EncryptionHelper:
-    """Helper class for encryption and decryption (caches key generation)"""
-
-    _cipher_suite = None
-
-    @classmethod
-    def get_cipher_suite(cls):
-        """Get Fernet cipher suite (singleton pattern)"""
-        if cls._cipher_suite is None:
-            secret_key = settings.SECRET_KEY.encode()[:32].ljust(32, b"0")
-            encryption_key = base64.urlsafe_b64encode(secret_key)
-            cls._cipher_suite = Fernet(encryption_key)
-        return cls._cipher_suite
-
-
-def encrypt_api_key(plain_text: str) -> str:
-    """Encrypt API key"""
-    cipher_suite = EncryptionHelper.get_cipher_suite()
-    return cipher_suite.encrypt(plain_text.encode()).decode()
-
-
-def decrypt_api_key(encrypted_text: str) -> str:
-    """Decrypt API key"""
-    cipher_suite = EncryptionHelper.get_cipher_suite()
-    return cipher_suite.decrypt(encrypted_text.encode()).decode()
-
-
-def is_encrypted(text: str) -> bool:
-    """Check if text is already encrypted"""
-    try:
-        cipher_suite = EncryptionHelper.get_cipher_suite()
-        cipher_suite.decrypt(text.encode())
-        return True
-    except Exception:
-        return False
